Set various -W flags on apple
Apple supports anything here that Clang supports and should have them
set the same as everywhere else.

Most importantly this gives apple the -Wno-deprecated-declarations flag
which has been driving me nuts on macos.

This also version-gates the -Wno-deprecated-declarations so that it
will turn on again when we bump the version beyond .10.

+++ llarp/apple/PacketTunnelProvider.m
@@ -91,7 +91,7 @@ del_ipv4_route(const char* addr, const char* netmask, void* ctx)
   LLARPPacketTunnel* t = (__bridge LLARPPacketTunnel*)ctx;
   NSMutableArray<NEIPv4Route*>* routes =
       [NSMutableArray arrayWithArray:t->settings.IPv4Settings.includedRoutes];
-  for (int i = 0; i < routes.count; i++)
+  for (size_t i = 0; i < routes.count; i++)
   {
     if ([routes[i].destinationAddress isEqualToString:route.destinationAddress] &&
         [routes[i].destinationSubnetMask isEqualToString:route.destinationSubnetMask])
@@ -137,7 +137,7 @@ del_ipv6_route(const char* addr, int prefix, void* ctx)
   LLARPPacketTunnel* t = (__bridge LLARPPacketTunnel*)ctx;
   NSMutableArray<NEIPv6Route*>* routes =
       [NSMutableArray arrayWithArray:t->settings.IPv6Settings.includedRoutes];
-  for (int i = 0; i < routes.count; i++)
+  for (size_t i = 0; i < routes.count; i++)
   {
     if ([routes[i].destinationAddress isEqualToString:route.destinationAddress] &&
         [routes[i].destinationNetworkPrefixLength
